backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudscraper==1.2.71
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
fake-useragent==2.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.40
langchain-ollama==0.2.3
langsmith==0.3.11
ollama==0.4.7
orjson==3.10.15
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zstandard==0.23.0
